+++ infra_ai_service/api/ai_enhance/text_process.py
@@ -1,37 +1,11 @@
-from fastapi import APIRouter, HTTPException
-from pydantic import BaseModel
-import re
-import logging
+from fastapi import APIRouter
 
-# 设置日志
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+from infra_ai_service.model.model import TextInput, TextOutput
+from infra_ai_service.service.text_service import process_text
 
 router = APIRouter()
 
 
-class TextInput(BaseModel):
-    content: str
-
-
-class TextOutput(BaseModel):
-    modified_content: str
-
-
-def clean_text(text: str) -> str:
-    # 正确转义正则表达式中的特殊字符
-    return re.sub(r'[{}\[\]()@.#\\_\':\/-]', '', text)
-
-
 @router.post("/process/", response_model=TextOutput)
-async def process_text(input_data: TextInput):
-    try:
-        modified_text = clean_text(input_data.content)
-        return TextOutput(modified_content=modified_text)
-    except re.error as e:
-        logger.error(f"Regex error: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=400, detail="Regex processing error")
-    except Exception as e:
-        logger.error(f"Error processing text: {str(e)}", exc_info=True)
-        raise HTTPException(status_code=400,
-                            detail=f"Error processing text: {str(e)}")
+async def process_text_api(input_data: TextInput):
+    return await process_text(input_data.content)
