+++ infra_ai_service/sdk/qdrant.py
@@ -1,0 +1,3 @@
+from infra_ai_service.api.common.utils import setup_qdrant_environment
+
+fastembed_model, qdrant_client, collection_name = setup_qdrant_environment()
